[#10] Start create engine

+++ RogaliqueGame/Constants.h
@@ -2,8 +2,8 @@
 #define PI 3.14159265
 
 namespace Rogalique {
-	const float SCREEN_WIDTH = 800.f;
-	const float SCREEN_HEIGHT = 600.f;
+	const int SCREEN_WIDTH = 800;
+	const int SCREEN_HEIGHT = 600;
 	const std::string PATH_TO_SAVE = "Save/\\save.txt";
 	const std::string GAME_NAME = "Rogalique Sus";
 }
